requirements.txt
Keras>=2.2.4
tensorflow-gpu=1.13.1
msgpack>=0.6.1
h5py>=2.9.0
